Bug fix on beer resource update
Fixed a bug causing beer update to not fully process and save. Had to
call on beer as object.

Removed handlebars for show single beer by ID. Instead modified how
beer was passed so showing a single w/ handlebars could be used by
the same handlebars passing an index of multiple beers into the UI - had
to wrap in brackets.