Modify sign-up and sign-up to a modal
Moved sign-up and sign-in forms fields and submit buttons to modals.
Added code in app.js to the submit event handler for sign-up and
sign-in to make the BS modal toggle off on submit.

Retained 'unauthenticated' class for these, so the buttons that
activate these modals will only be available to a logged out user.