Create beer style drop-down
Added available beer styles to a drop-down list in the 'create new beer'
modal feature.
List includes all major styles and some popular sub-styles. Any missing
options can be added later, or added in the description field.
Made it so placeholder selection on dropdown list is not a valid option
to select and if field is left blank will not log the style (style is
an optional field entirely).